광산 수입/지출 구조와 로그구현
- 광산의 수입을 델리게이트로 구현해서 게이지가 한 바퀴 돌 때마다 호출
- 로그를 찍기 위해서 맵 설정 수정
- 광산 업그레이드 기능 구현
- TODO : 비활성화 광산 그리드 방식으로 배치 예정

+++ IdleSimulation/Actor/Player.h
@@ -13,13 +13,18 @@ class Player : public Actor
 public:
 	Player() : gold(0) {}
 
+	// 내가 가진 자산이 광산의 구입비용보다 많거나 같은지 비교
 	bool TryPurchase(int cost);
+
+	// 내가 소유한 광산 목록에 추가
 	void AddMineList(Mine* mine) { ownedMines.emplace_back(mine); }
+
+	// 내가 소유한 광산 목록 꺼내기
 	const std::vector<Mine*>& GetOwnedMines() const { return ownedMines; }
 
 	// 자산 관리
 	long long GetGold() const { return gold; }
-	void AddGold(int amount) { gold += amount; }
+	void AddGold(long long amount) { gold += amount; }
 
 private:
 	// 소유한 자산
